Moved Processor into its own scope/namespace. Hugely untested of course.
